+++ R/proj_equidistant.R
@@ -1,6 +1,6 @@
-#' Projection for the area of interest with equal area projection
+#' Projection for the area of interest with equidistant projection
 #'
-#' Auto selecting equal-area projections based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected.\cr
+#' Auto selecting equidistant projections based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected.
 #' Please note that there's a longitudinal (less than 160) and latitudinal (less than 80) range limit. If area of interest larger than this please use [proj_hemisphere()] or use [proj_specify()].
 #' @param obj Input geo data, should be one of:\cr
 #'  - An object can be accepted by [sf::st_bbox()] to compute the bounding box\cr
@@ -11,10 +11,10 @@
 #'
 #' @returns A `proj4` or `WKT` string
 #' @export
-#' @seealso [proj_region()], [proj_conformal()], [proj_equidistant()]
+#' @seealso [proj_region()], [proj_equal_area()], [proj_conformal()]
 #'
-#' @examples proj_equal_area(c(xmax=112,xmin=156,ymin=6,ymax=23))
-proj_equal_area <- function(obj,output_type = "proj4",datum = "WGS84", unit = "m") {
+#' @examples proj_equidistant(c(xmax=112,xmin=156,ymin=6,ymax=23))
+proj_equidistant <- function(obj,output_type = "proj4",datum = "WGS84", unit = "m") {
   if (!(is.vector(obj) & identical(sort(names(obj)), sort(c("xmin", "xmax", "ymin","ymax"))))) {
     if(!sf::st_is_longlat(obj)) {
       obj = sf::st_transform(obj, 4326)
@@ -54,26 +54,47 @@ proj_equal_area <- function(obj,output_type = "proj4",datum = "WGS84", unit = "m
 
   # distance check
   lonlat_m = check_lonlat_dis(latmin, latmax, dlon)
-
+  ratio <- lonlat_m$dlat_m / lonlat_m$dlon_m
   if (max(lonlat_m$dlat_m, lonlat_m$dlon_m)  < 1e6) {
     message("## The map extent is not quite large")
-    message("## Select Lambert azimuthal equal area projection")
-    outputTEXT <- stringLinks("laea", NaN, center$lat, NaN, NaN, center$lng, NaN, datum, unit)
+    message("## Select Oblique azimuthal equidistant projection")
+    outputTEXT <- stringLinks("aeqd", NaN, center$lat, NaN, NaN, center$lng, NaN, datum, unit)
   } else {
     message("## The map extent is relatively large, choose projection considering map shape")
-    # ratio check
-    ratio <- lonlat_m$dlat_m / lonlat_m$dlon_m
-    if (ratio > 1.25) {
+    if (abs(center$lat) > 70) {
+      message("## Close to poles")
+      # case: close to poles
+      outputTEXT <- stringLinks("aeqd", NaN, sign(center$lat) * 90.0, NaN, NaN, center$lng, NaN, datum, unit)
+    }  else if (ratio > 1.25) {
       message("## North-south extent")
-      outputTEXT <- printNSextent("Equalarea", center,latmax,latmin, datum, unit)
-    } else if (ratio < 0.8) {
-      message("## East-west extent")
-      outputTEXT <- printEWextent("Equalarea", center,latmax,latmin,dlon, datum, unit)
+      # case: with an north-south extent
+      outputTEXT <- stringLinks("cass", NaN, NaN, NaN, NaN, center$lng, NaN, datum, unit)
+    } else if (abs(center$lat) < 15) {
+      message("## Close to equator")
+      # case: close to equator
+      if ((latmax * latmin) <= 0 ) {
+        message("## Extent is touching or crossing equator")
+        latS = max(abs(latmax), abs(latmin)) / 2
+      } else{
+        message("## Extent is not crossing equator")
+        latS = center$lat
+      }
+      outputTEXT <- stringLinks("eqc", NaN, NaN, latS, NaN, center$lng, NaN, datum, unit)
     } else {
-      message("## Square-shaped extent")
-      outputTEXT <- printSquareFormat("Equalarea", center,latmax,latmin, datum, unit)
+      message("## Mid-Latitude away from pole and equator")
+      # case: between pole and equator
+      # computing standard paralles
+      # interval <- (latmax - latmin) / 6
+      # Oblique azimuthal equidistant
+      message("## Select Oblique azimuthal equidistant projection")
+      outputTEXT <- stringLinks("aeqd", NaN, center$lat, NaN, NaN, center$lng, NaN, datum, unit)
+      # outputTEXT <- list(
+      #   "Equidistant conic" = stringLinks("eqdc", NaN, center$lat, latmin + interval, latmax - interval, center$lng, NaN),
+      #   "Oblique azimuthal equidistant" = stringLinks("aeqd", NaN, center$lat, NaN, NaN, center$lng, NaN)
+      # )
     }
   }
+
 
   if(output_type == "proj4") {
     return(outputTEXT$PROJ)
@@ -82,7 +103,5 @@ proj_equal_area <- function(obj,output_type = "proj4",datum = "WGS84", unit = "m
   }
 
 }
-
-