+++ tests/testthat/test-proj_region.R
@@ -7,8 +7,7 @@ test_that("proj_region Equalarea works", {
                  proj_region("Equalarea"),"+proj=tcea +lon_0=28 +datum=WGS84 +units=m +no_defs")
   expect_equal(sf::st_bbox(c(xmin = 13,ymin = -53,xmax = 80 ,ymax = -25),crs = 4326) |>
                  proj_region("Equalarea"),"+proj=aea +lon_0=46.5 +lat_1=-48.3333333 +lat_2=-29.6666667 +lat_0=-39 +datum=WGS84 +units=m +no_defs")
-  expect_equal(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326) |>
-                 proj_region("Equalarea"),"+proj=cea +lon_0=86 +lat_ts=-12 +datum=WGS84 +units=m +no_defs")
+  expect_error(proj_region(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326) ,"Equalarea"))
   expect_equal(sf::st_bbox(c(xmin = 168,ymin = 40,xmax = -171 ,ymax = 55),crs = 4326) |>
                  proj_region("Equalarea"),"+proj=laea +lon_0=178.5 +lat_0=47.5 +datum=WGS84 +units=m +no_defs")
   expect_equal(proj_region(sf::st_bbox(c(xmin = -87,ymin = -4,xmax = -71 ,ymax = 10),crs = 4326),property = "Equalarea"),
@@ -45,8 +44,7 @@ test_that("proj_region Conformal works", {
                  proj_region("Conformal"),"+proj=tmerc +lon_0=28 +datum=WGS84 +units=m +no_defs")
   expect_equal(sf::st_bbox(c(xmin = 13,ymin = -53,xmax = 80 ,ymax = -25),crs = 4326) |>
                  proj_region("Conformal"),"+proj=lcc +lon_0=46.5 +lat_1=-48.3333333 +lat_2=-29.6666667 +lat_0=-39 +datum=WGS84 +units=m +no_defs")
-  expect_equal(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326) |>
-                 proj_region("Conformal"),"+proj=merc +lon_0=86 +lat_ts=-12 +datum=WGS84 +units=m +no_defs")
+  expect_error(proj_region(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326),"Conformal"))
   expect_equal(sf::st_bbox(c(xmin = 160,ymin = 43,xmax = -148,ymax = 56),crs = 4326) |>
                  proj_region("Conformal"),"+proj=lcc +lon_0=-174 +lat_1=45.1666667 +lat_2=53.8333333 +lat_0=49.5 +datum=WGS84 +units=m +no_defs")
   expect_equal(sf::st_bbox(c(xmin = 146,ymin = 24,xmax = -174,ymax = 55),crs = 4326) |>
@@ -71,16 +69,24 @@ test_that("proj_region Conformal works", {
  UNIT["Meter",1.0]]'
   )
   expect_equal(proj_region(sf::st_bbox(c(xmin = 80,ymin = 84,xmax = 91 ,ymax = 85),crs = 4326),property = "Conformal"),
-               "+proj=stere +lon_0=85.5 +lat_0=90 +k_0=0.994 +datum=WGS84 +units=m +no_defs")
+               "+proj=stere +lon_0=85.5 +lat_0=84.5 +datum=WGS84 +units=m +no_defs")
   expect_equal(proj_region(sf::st_bbox(c(xmin = 38,ymin = -83,xmax = 110 ,ymax = -80),crs = 4326),property = "Conformal"),
-               "+proj=stere +lon_0=74 +lat_0=-90 +k_0=0.994 +datum=WGS84 +units=m +no_defs")
+               "+proj=stere +lon_0=74 +lat_0=-90 +datum=WGS84 +units=m +no_defs")
   expect_equal(proj_region(sf::st_bbox(c(xmin = 106,ymin = 28,xmax = 107 ,ymax = 31),crs = 4326),property = "Conformal"),
-               "+proj=tmerc +x_0=5e+05 +lon_0=106.5 +k_0=0.9999 +datum=WGS84 +units=m +no_defs")
+               "+proj=stere +lon_0=106.5 +lat_0=29.5 +datum=WGS84 +units=m +no_defs")
   expect_equal(proj_region(sf::st_bbox(c(xmin = 106,ymin = 28,xmax = 110 ,ymax = 30),crs = 4326),property = "Conformal"),
-               "+proj=tmerc +x_0=5e+05 +lon_0=108 +k_0=0.9996 +datum=WGS84 +units=m +no_defs")
+               "+proj=stere +lon_0=108 +lat_0=29 +datum=WGS84 +units=m +no_defs")
   expect_error(proj_region(c(xmin = 2,ymin = -23,xmax = 190 ,ymax = -1),"Equidistant"))
   expect_equal(proj_region(list(xmin = 106,ymin = 28,xmax = 110 ,ymax = 30),property = "Conformal"),
-               "+proj=tmerc +x_0=5e+05 +lon_0=108 +k_0=0.9996 +datum=WGS84 +units=m +no_defs")
+               "+proj=stere +lon_0=108 +lat_0=29 +datum=WGS84 +units=m +no_defs")
+  expect_equal(proj_region(c(xmin = 5,ymin = 12,xmax = 7 ,ymax = 13),property = "Conformal"),
+               "+proj=stere +lon_0=6 +lat_0=12.5 +datum=WGS84 +units=m +no_defs")
+  expect_equal(proj_region(c(xmin = 5,ymin = 12,xmax = 7 ,ymax = 13),property = "Conformal"),
+               "+proj=stere +lon_0=6 +lat_0=12.5 +datum=WGS84 +units=m +no_defs")
+  expect_equal(proj_region(c(xmin = 5,ymin = 12,xmax = 7 ,ymax = 30),property = "Conformal"),
+               "+proj=tmerc +x_0=5e+05 +lon_0=6 +k_0=0.9999 +datum=WGS84 +units=m +no_defs")
+  expect_equal(proj_region(c(xmin = 7,ymin = 12,xmax = 11 ,ymax = 30),property = "Conformal"),
+               "+proj=tmerc +x_0=5e+05 +lon_0=9 +k_0=0.9996 +datum=WGS84 +units=m +no_defs")
 })
 
 
@@ -91,10 +97,10 @@ test_that("proj_region Equidistant works", {
                  "+proj=cass +lon_0=28 +datum=WGS84 +units=m +no_defs")
   expect_equal(proj_region(sf::st_bbox(c(xmin = 13,ymin = -53,xmax = 80 ,ymax = -25),crs = 4326),"Equidistant"),
                  "+proj=aeqd +lon_0=46.5 +lat_0=-39 +datum=WGS84 +units=m +no_defs")
-  expect_equal(proj_region(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326),"Equidistant"),
-                "+proj=eqc +lon_0=86 +lat_ts=-12 +datum=WGS84 +units=m +no_defs")
-  expect_equal(proj_region(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),"Equidistant"),
-               "+proj=eqc +lon_0=86 +lat_ts=-12 +datum=WGS84 +units=m +no_defs")
+  expect_error(proj_region(sf::st_bbox(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),crs = 4326),"Equidistant"))
+
+  expect_error(proj_region(c(xmin = 2,ymin = -23,xmax = 170 ,ymax = -1),"Equidistant"))
+
   expect_equal(proj_region(sf::st_bbox(c(xmin = 142,ymin = 25,xmax = -143 ,ymax = 32),crs = 4326),"Equidistant"),
                "+proj=aeqd +lon_0=179.5 +lat_0=28.5 +datum=WGS84 +units=m +no_defs")
 })